relax mutex in invoke call
This fixes large lag when waiting on mutex, when
user requested data/action that doesn't change but
another action that requires mutex is pending.
This fixes access delay when switching wallets in
Cake Wallet.

We were making multiple invoke() calls to obtain
wallet name, decrypt a string to get passphrase
and finally get the encrypted seed, now we do
exactly the same but we don't have to wait for
other calls to finish.